Skybox Shader Complete + Variable Name Change
Completed the skybox shader (Unable to test as KodeLife does not include functionality for samplerCube so will need to test in-engine). Changed the variable names in each project and code file to be the same in each file. Used numbers instead of letters so they could more easily be looped through