Add in imgui tea 31 (#19)
* Added Basic GUI Support

Added the ability to create a window. Shall be building out functionality

* Updates to Names + Variables

Updated the name of the files to be more descriptive

Added a Width and Height variable to CreateWindow() so you can set the size you want by default

* Added AddButton + Split Create and End Window

Added the ability to create a button with a given name, width and height

Split the Start and End Window Creation into two functions. Allows us to add things between these two functions before Window Creation Ends

* Added SetPosition

Allows for the user to set the initial position that an element will be drawn from

* Added Checkbox + Changed Names to Strings

Added the ability to add checkboxes to the window

Changed names to use strings instead of c_strings (Although are still converted to c_strings to ImGui)

* Added Sliders + Trying to Fix Bug

Added an Integer and Float Slider

Trying to fix a bug where the position and size are set permanently. I know this is because its being reset every frame, just trying to figure out a way to cleanly get it to run once

* Fixed Bug

Fixed the Bug where it would reset the position and size of the window after moving/resizing. Added ImGuiCond_Once condition to the initial position and size funcitons

* Added Window Flags parameter

Added a parameter to store Window Flags for later setup

* Added Vector Sliders for Integers & Floats

Added in sliders. Took a while as I wasn't happy with splitting each vector size into its own function, so combined them all together.

* Trying to switch to Smart Pointers

Trying to use Smart Pointers instead of Raw Pointers but having some issues updating values

* Added Back initial changes

Added back changes asked for by team

* Added back OBJ Model files + Unignored Assets

* Changed Errors to use VK_CRITICAL

* Update .gitignore

Changed the folder unignore for assets to be more general, so as not to tie it into Vakol specifically

* Updated gitignore

Updated to add a line for coreAssets folder

* Update CMakeLists.txt

Changed 'option' to 'set' to fix bug with ASSIMP not finding the obj importer

* Removed raw pointers

Removed raw pointers being passed around from function to function

* Update Application.cpp

* Update Application.cpp

Removed ImGUI test strings