Terrain foliage tea 146 (#44)
* fixing performance issues

* added in fly cam

* Uploaded new Heightmap

Uploaded a new heightmap for the terrain in different resolutions. Removed the older ones

* Added CC0 Tree Asset + Added Instancing to lua

Added CC0 Tree Asset from OpenGameArt with the name of the creator included for referencing purposes

Added Instancing to lua by tying the function from Instance.cpp into LuaAccess.cpp

* Moved 'CreateInstances'

Moved 'CreateInstances' to RegisterOther so there will be less function calls

* Adding in instanced trees

Trying to get instanced trees to work, having a few issues

* Update LuaAccess.cpp

Trees are now displaying (Cheers Caleb & James for the help)

* Update instancing.lua

Having trouble getting multiple trees to show. Can get one, but not more

* yeah

* before i forget

* big blend

* the kiki army goes stronger every minute

* kiki army is TOO STRONG

* tired

* Delete instancing.lua

* Created Trees for Terrain

* Selectable Backface Culling

When adding a model in lua, you can now select whether that model has backface culling applied or not. This will allow you to have backface culling applied or unapplied to that model specifically

* Instancing Buggy

Instancing seems to work with one model but not with multiple at once

* Instancing Working (Thanks Caleb)

Got instancing working (Thanks Caleb for the help). Able to get around 130-160 FPS depending on the direction I look in, and that's with 2000 rocks and 6000 trees

---------

Co-authored-by: Wolfleader <james.morgan@iinet.net.au>
Co-authored-by: Caleb <cjmeepers@gmail.com>