Added Centering for ImGui Windows
Added the ability to center the Dear ImGui windows when creating them. You can select to center on the X and Y Axis, and apply an offset to the values if you wish to offset the centered position. If the size of the window is 0 on either the X or Y, the centering will be turned off for that axis regardless of the boolean value, as there is no way to determine what the size of the window is until content is added