Hud with player stats tea 156 (#47)
* Added Default Window Style + Text Colour Option

Added Default Window Style so when we start the creation of a window, it will default to that style

Added Text Colour Option so when we create text, we can give it an RGB colour and opacity

* Added Player Stats + Fixed Text Centering

Added Player Stats functions to increment and decrement stats by input amounts

Fixed Text Centering, it wasn't taking into account the scaled size of the text

* Added lua calls for window styling

Added lua calls for window styling (Changing the background colour and changing the background edge rounding)

* Added Centering for ImGui Windows

Added the ability to center the Dear ImGui windows when creating them. You can select to center on the X and Y Axis, and apply an offset to the values if you wish to offset the centered position. If the size of the window is 0 on either the X or Y, the centering will be turned off for that axis regardless of the boolean value, as there is no way to determine what the size of the window is until content is added

* Added the ability to add content to the same line

Added the ability to add content to the same line so we can put content next to each other

* Added Display Window Size Gets + Added Icons

Added Display Window Size Get functions which return the width and height of the current display window. This means we can have things be relative to the size of the window

Added test icons for the health, hunger and thirst

* Update start_menu.lua

* Update start_menu.lua

I have set a minimum size for the window as a test in start_menu. If the value of 'GUI:get_display_window_width() / 3' falls below the value of 'WINDOWWIDTHMINIMUM', it will default to 'WINDOWWIDTHMINIMUM'

* Update start_menu.lua

Set a Window Width Minimum and Maximum, so now the window will not exceed a certain size or go below a certain size

* Update start_menu.lua

* Moved Icons to Demo Assets

Moved Icons to Demo Assets as they are a part of the demo and not core assets